Sales Analysis dashboard using superstore dataset
This report is dedicated to advancing my proficiency in Tableau by creating a diverse range of visualizations. It includes the development of intricate visualizations such as heat maps for spatial data representation, pie charts for categorical analysis, funnel graphs to depict sequential processes, and conditional formatting to highlight data trends effectively. Additionally, it encompasses the creation of line and bar charts for trend analysis and comparison, dual-axis charts to visualize multiple measures simultaneously, and the use of reference bands for benchmarking and target analysis. Furthermore, the report incorporates state-wise mapping to geographically analyze data trends and patterns across different regions.